chore(user): Rename query.js
- Rename query.js to query.test.js
- update all imports

[Delivers 164776738]
